Created/Implemented Intro Video
I created a short intro video explaining the premise of the game and implemented that video as the first scene to load when the player selects "Start Game" from the main menu. I have also updated the teleporter in Level 0 to teleport the player to Scene 4 from it's previous setting of Scene 3, to account for the addition of the Intro to the scene list.